Project (For Mqtt)/requirements.txt
paho-mqtt
requests
python-dotenv